+++ src/parser/api/source.test.ts
@@ -16,6 +16,7 @@ describe('Unit: parser/api/source', () => {
       assert(breaklines('<# a #>b\nc') === '<# a #>b\\\nc');
       assert(breaklines('<# a\nb #>c\nd') === '<# a\nb #>c\\\nd');
       assert(breaklines('<# a\nb\nc #>d\ne') === '<# a\nb\nc #>d\\\ne');
+      assert(breaklines('((a\nb))c') === '((a\nb))c');
       assert(breaklines('>a') === '>a');
       assert(breaklines('>a\nb') === '>a\nb');
       assert(breaklines('a\n\nb\n') === 'a\n\nb\n');